+++ config/dbConstants.py
@@ -8,3 +8,6 @@ SECT_LOG = 'LOGGERS'
 DB_IO = 'db_input_output'
 DB_COMMANDS = 'db_commands'
 CONF_HANDLER = 'config_handler'
+
+VALID_LOG_LEVELS = ['DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL']
+CONFIG_FILE = 'config/config.ini'